Refactor: consts, separate file for errors
